Travel
I added a travel page. I decided to add this page to document my travels in the upcoming semester. 

The largest challenge of this page was implementing a google maps window with a pin in my home base, San Sebastian Spain.  I did not produce complicated code, but this was my first experience using JavaScript and using an API.
--travel.html
----the html layout of the page
--custom.css
----a short css script detailing the Google Maps window's format
--custom.js
----a short JavaScript script initializing the window